Actualizacion gramatica - 4. Hecho: - Agregadas reglas para generar la polaca inversa en asignaciones, ifs y loops. - Actualizada accion semantica GeneraTokenParticular para poder agregar a la polaca la representacion (en string) del token en cuestion. - Cambiados el tipo de cada token de int a short, para facilitar uso del parser. - Acomodada la clase Compilador para que permita activar o desactivar los prints de los datos.
Falta hacer:
- Agregar regla para detectar error cuando falta parentesis de cierre para parametros (invocacion).
- Agregar regla para detectar error cuando faltan comas para separar lista variables.
- Agregar regla para detectar error cuando falta la palabra clave END_IF.
- Generacion de representacion en polaca inversa para invocaciones, out y procedimientos.
- Incorporar mas datos necesarios a la TS (tipo, uso, ambito, etc).
- Chequeos semanticos para redeclaracion de variables y declaracion inexistente.
- Chequeos para evitar operaciones entre distintos tipos de variables.
- Chequeos semanticos en invocacion de procedimientos (NI, n° params, tipo params).

+++ src/main/java/analizador_lexico/acciones_semanticas/GeneraTokenDouble.java
@@ -17,9 +17,9 @@ public class GeneraTokenDouble extends AccionSemantica {
 
     private final TablaSimbolos tablaS;
 
-    private final int token;
+    private final short token;
 
-    public GeneraTokenDouble(MaquinaEstados maquinaEstados, TablaSimbolos tablaS, int token) {
+    public GeneraTokenDouble(MaquinaEstados maquinaEstados, TablaSimbolos tablaS, short token) {
         this.maquinaEstados = maquinaEstados;
         this.tablaS = tablaS;
         this.token = token;
