+++ src/main/java/analizador_lexico/acciones_semanticas/GeneraTokenUINT.java
@@ -14,9 +14,9 @@ public class GeneraTokenUINT extends AccionSemantica {
 
     private final TablaSimbolos tablaS;
 
-    private final int token;
+    private final short token;
 
-    public GeneraTokenUINT(MaquinaEstados maquinaEstados, TablaSimbolos tablaS, int token) {
+    public GeneraTokenUINT(MaquinaEstados maquinaEstados, TablaSimbolos tablaS, short token) {
         this.maquinaEstados = maquinaEstados;
         this.tablaS = tablaS;
         this.token = token;
